Make it easier to add more devices
* Move src/hantek/usb to src/usb
* Move ControlBegin and ControlGetSpeed to src/usb as they are the only protocol structs that are required to make the usb module independently compilable.
* Move hantek protocol specifc structs and enums to src/hantekprotocol
* Rename src/hantek to src/hantekdso
* Refactor all model initalisation specifc code out of hantekdsocontrol into own subclasses inheriting from DSOModel
* Split definitions.h file into controlindexes.h, controlspecification.h, enums.h, states.h

+++ openhantek/src/configdialog/DsoConfigAnalysisPage.cpp
@@ -15,7 +15,7 @@ DsoConfigAnalysisPage::DsoConfigAnalysisPage(DsoSettings *settings, QWidget *par
     windowFunctionLabel = new QLabel(tr("Window function"));
     windowFunctionComboBox = new QComboBox();
     windowFunctionComboBox->addItems(windowFunctionStrings);
-    windowFunctionComboBox->setCurrentIndex(settings->scope.spectrumWindow);
+    windowFunctionComboBox->setCurrentIndex((int)settings->scope.spectrumWindow);
 
     referenceLevelLabel = new QLabel(tr("Reference level"));
     referenceLevelSpinBox = new QDoubleSpinBox();
